Move config values out of static mutexes.
The behaviour of this should be identical to the previous behaviour, but with less mutex overhead, which isn't neccessary when accessing immutable data.

This still contains a bunch of statics that could be replaced with similar variables, especially considering connection pooling, but that should be done in a later patch.